mlops/aml-requirements.txt
# required for job submission
# pinned to version where required features added to ScriptRunConfig
azureml-core>=1.20.0
azureml-mlflow>=1.20.0
azureml-train>=1.20.0
azureml-dataprep>=1.20.0
azureml-widgets>=1.20.0
pyJWT==1.7.1
numpy==1.17.5
tensorflow==2.3.1